+++ Source/Util/Dictionary+Update.swift
@@ -1,11 +1,11 @@
-// MIT license. Copyright (c) 2020 SwiftyFORM. All rights reserved.
+// MIT license. Copyright (c) 2021 SwiftyFORM. All rights reserved.
 import Foundation
 
 extension Dictionary {
 	/// merge two dictionaries into one dictionary
 	mutating func update(_ other: Dictionary) {
 		for (key, value) in other {
-			self.updateValue(value, forKey:key)
+			self.updateValue(value, forKey: key)
 		}
 	}
 }