+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 ## 📄 Change Log
 
+### 1.8.6
+
+ * Contributions by Luca Pizzini. Thank you.
+   The background color is now customizable for buttons.
+   Extended the README with guidance for how to install SwiftyFORM manually.
+ * Renamed from `CustomizableLabels` to `CustomizableLabel`, so it's consistent with the new `CustomizableButton` protocol.
+ * Fixed broken PrecisionSliderFormItem. With iOS14 Apple have changed the view hierarchy inside UITableView.
+ * Example app: Renamed from `WorkInProgressViewController` to `LabViewController`.
+
 ### 1.8.5
 
  * Contributions by IOS-Sisa. Thank you.